+++ src/SnippetsManager.js
@@ -57,34 +57,40 @@ class SnippetsManager {
             }
 
             if (snippets[snippet.name] !== undefined) {
-                vscode.window.showErrorMessage(`A snippet '${snippet.name}' already exists - so adding a unique id`);
-                // return;
+                let proposedName = snippet.name
                 snippet.name = snippet.name + '_' + this.uuidv4()
+                vscode.window.showErrorMessage(`A snippet '${proposedName}' already exists - so adding a unique id: '${snippet.name}'`);
             }
 
+            // if user has not entered anything for 'prefix' or 'description', auto set to some sensible values
             if (snippet.prefix == "")
                 snippet.prefix = snippet.name;
             if (snippet.description == "")
-                snippet.description = snippet.name + ' description';
+                snippet.description = `description for ${snippet.name}`;
 
-            let edit = jsonc.modify(jsonText, [snippet.name],
-                {
-                    prefix: snippet.prefix,
-                    body: snippet.body,
-                    description: snippet.description
-                },
-                {
-                    formattingOptions: {
-                        // based on default vscode snippet format
-                        tabSize: 2,
-                        insertSpaces: false,
-                        eol: ''
-                    }
-                });
+            // Edit the snippet json file by inserting the snippet entry
+            let path = snippet.name
+            let snippetEntry = {
+                prefix: snippet.prefix,
+                body: snippet.body,  // array of strings
+                description: snippet.description
+            }
+            let options = {
+                formattingOptions: {
+                    // If indentation is based on spaces (`insertSpaces` = true), then what is the number of spaces that make an indent?
+                    tabSize: 2,  // based on default vscode snippet format
 
-            let fileContent = jsonc.applyEdits(jsonText, edit);
+                    // Is indentation based on spaces?
+                    insertSpaces: false,
+
+                    // The default 'end of line' character
+                    eol: ''
+                }
+            }
+            let edit = jsonc.modify(jsonText, [path], snippetEntry, options); // create the 'edit'
+            let fileContent = jsonc.applyEdits(jsonText, edit); // apply the 'edit'
             fs.writeFile(snippetFile, fileContent, () => { });
-            vscode.window.showInformationMessage(`Snippet '${snippet.prefix}' added to ${snippet.language} snippets`)
+            vscode.window.showInformationMessage(`Snippet '${snippet.prefix}' added to '${snippet.language}.json' snippets. Tip: Edit snippets using the command "Configure User Snippets"`)
         });
 
     }